Removes call to .lower on vnf_member_index in alarm creation
Currently this does not produce an error just because vnf_member_index
is been passed as a string in the kafka msg. Considering that in the IM
it is a integer, this may change in the future, in which case this would
fail. It also does not make any sense and can be considered a typo.

Signed-off-by: Benjamin Diaz <bdiaz@whitestack.com>