Replaces vdu-id-ref for name during VDU lookup
Vdus were being looked up using the vdu-id-ref param, which
corresponds to the name used in the vnfd to define that vdu.
That was not the intended behaviour, because that param is clearly
not unique. It has been replaced with the name param, which
is unique to each vdu record.

Signed-off-by: Benjamin Diaz <bdiaz@whitestack.com>