Reduce Resource Consumption
The vmware vnf collector requests a time series of data from
vROPS and then parses the whole block in order to retrieve the
most recent data element.  The time frame for the request
should be limited to a sane value, such as 1 hour.

The look ups for items that do not change should be outside
of the loop for looking up the metrics.

Add more metrics to list of supported metrics.

Improve error handling.

BUG 784

Change-Id: Ie3cef457eaf657c18ba682e3c5d3abf40a759e69
Signed-off-by: beierl <mbeierl@vmware.com>