Fix MON Stein Metrics
Gnocchi deprecated the use of the cpu_util parameter after Rocky.  What
is needed instead is to use the ceilometer-low-rate cpu metric and
covert that to a percentage using value/time interval.  This metric
was introduced in Mitaka, and has been shown to work with Rocky as
part of the Whitestack WhiteCloud Rocky deployment, as well as Stein
with Canonical Charmed OpenStack Stein deployment.

Bug 790

Change-Id: I1a895e3569f9206c39ce3c77636d1b45a1a1bd3a
Signed-off-by: beierlm <mark.beierl@canonical.com>